Moved Twig extension to ContentBundle
